review fixes: diff-review findings (robustness + 2 UX gaps)
From the v0.0.4..ratdb_fixes review (adversarially verified):

- web: retry RatDB load after a transient db_init failure instead
  of caching the failure for the session (dbLoadAttempts.delete on
  catch) -- previously one failed fetch disabled RatDB until reload.
- web: sync the share URL after a closing commit rewrites the
  angle input to canonical form (programmatic .value fires no input
  event, so ?seq= was left stale).
- web: .catch() the fire-and-forget augmentInfoWithDbId so a
  wasm-side rejection degrades to 'no RatDB row' instead of an
  unhandled rejection.
- db_seq_of: reject non-finite/negative f64 ids before the u64
  cast (would saturate to index 0 and return a wrong rat); guards
  the public wasm export beyond the JS call sites.
- lex_rank_of: accumulate the rank in u64 (was usize: 32-bit on
  wasm32) and return Option<u64>, matching the lazy reader; drops
  the now-redundant cast in RatDafsa::index_of.
- blocks_schema.txt: disambiguate the Capacity note's n_edges
  (manifest total u64 vs per-block header u32).

(One reported item left as-is: previewAngle=0 after a non-closing
commit is the intended 'arm the next straight-ahead preview', same
as the pre-refactor keyboard path.)

+++ src/stringmatch/dafsa/core.rs
@@ -288,10 +288,12 @@ impl Dafsa {
     /// state, the count of sequences in earlier-lex subtrees plus
     /// the empty continuation if the state itself is accepting.
     /// O(seq.len() * fanout).
-    pub fn lex_rank_of<S: AsRef<[i8]>>(&self, seq: S) -> Option<usize> {
+    pub fn lex_rank_of<S: AsRef<[i8]>>(&self, seq: S) -> Option<u64> {
         let seq = seq.as_ref();
         let mut s = 0u32;
-        let mut rank: usize = 0;
+        // u64 accumulator: a rank can exceed u32::MAX (~ZZ12 n=17),
+        // and `usize` is 32-bit on wasm32. Matches the lazy reader.
+        let mut rank: u64 = 0;
         for &c in seq {
             if self.states[s as usize].is_accept {
                 rank += 1;
@@ -300,7 +302,7 @@ impl Dafsa {
             let mut next: Option<u32> = None;
             for e in edges {
                 match e.label.cmp(&c) {
-                    std::cmp::Ordering::Less => rank += self.counts[e.target as usize],
+                    std::cmp::Ordering::Less => rank += self.counts[e.target as usize] as u64,
                     std::cmp::Ordering::Equal => {
                         next = Some(e.target);
                         break;
